[wasm][memory64] Fix and re-enable trap handling
Memory64 trap handling used the memory maximum_pages as the size of the
memory reserved for committed pages and guard pages. That did not work
well in two cases: when a module imports a memory that declares a
smaller maximum_pages and when memory allocation fails and
BackingStore::AllocateWasmMemory retries and reserves less memory than
the declared maximum_pages.

To fix these two issues, when trap handling is enabled we always
reserve kMaxMemory64Size (=16GB) for all 64-bit memories and when such
an allocation fails, we fail without retrying with smaller allocations.

Fixed: 388290793
Change-Id: I0554ca84804effa8211ea093a04db1640d9d6460
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6169203
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Commit-Queue: Paolo Severini <paolosev@microsoft.com>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98181}