[codegen][x64] Improve code for ResolveWasmCodePointer
Use a single comparison and a single conditional jump instead of two
32-bit comparisons and two jumps.
If the signature hash fits in a signed 32-bit integer, we just encode it
as an immediate (which is sign-extended on the comparison). Otherwise
load the 64-bit constant into a register first and then compare against
that.

Note that many signature hashes actually fit in a 32-bit signed integer
(all signatures with <=1 tagged return value and no return values on the
stack).

R=jkummerow@chromium.org

Bug: 389879224
Change-Id: I7bac7da25b1115371dc424a20068d3e21b2b81d8
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6172288
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98176}