[wasm] Add padding words for stack allocated param buffer in js-to-wasm
The param buffer of the generic js-to-wasm wrapper contains 3 regions,
one region for gp register parameters, one region for fp register
parameters, and one region for stack parameters. On arm and ia32, the
region for gp register parameters is 3 * 4 bytes big, plus 4 bytes for
padding so that the fp register parameter region is 8 byte aligned.
In total, the gp register parameter region is therefore 2 * 8 bytes big.

Since there are only 3 registers for parameters, only a single I64
parameter fits completely into registers, all other I64 parameters will
at least partly be in registers.

For the param buffer size this means that if there are 10 I64
parameters, only a single parameter fits into the 2 * 8 bytes region for
gp register parameters. 8 * 8 bytes will be unused, as there are no
parameters stored in fp registers, and 9 parameters use 9 * 8 bytes in
the stack region (half of one parameter is stored in a register, but the
other half on the stack needs padding). Therefore in total the buffer
size has to be 19 * 8 bytes, but so far we only reserved 18 * 8 bytes.

Bug: 390035515
Change-Id: Iaea5654ba4c6a93b77391697d6c9c09c86406b13
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6176748
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98152}