[wasm-c-api] Release adopted vector of results
This patch fixes a bug that slipped through the checks of a previous PR.
When performing a `wasm_func_call` we `adopt` the vector of results the
caller passes and will receive back with the results of the call;
however, that vector is never released, which can result in corrupted
data being passed back to the caller. Here we add a call to
`release_val_vec` before finishing the call.

Change-Id: I60ac8c2db218fd08351b28bd890b2fb15eca7393
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6182561
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Reviewed-by: Daniel Lehmann <dlehmann@chromium.org>
Auto-Submit: Edoardo Marangoni <edoardo@wasmer.io>
Commit-Queue: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98197}