Add trace_id param to tick-sample
It will be used to associate a trace event with a sample, when a sample
is taken via V8::CollectSample. This is a faster alternative to
collecting a stack trace and adding it to the trace event payload, since
this requires symbolizing the stack. Instead, V8::CollectSample only
collects the stack register data and leaves the symbolization to the
profiler thread running the SamplingEventsProcessor.

Note this change doesn't do anything with the trace id yet. A follow up CL will add this data to the traced CPU profile.

Bug: 390155857
Change-Id: Iffa492de02ee0cfb7fd1bb5daf5762b1309e6681
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6179435
Reviewed-by: Simon Zünd <szuend@chromium.org>
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>
Commit-Queue: Andres Olivares <andoli@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98237}