[v8/memory-reducer] Avoid rescheduling MemoryReducer while frozen
Currently, we sometimes run a GC after we have compacted Chrome's memory, which is counterproductive because we are immediately bringing back memory that we have just swapped out.

This CL aims to fix this by limiting when we can run a GC in v8, to prevent this from happening.

Bug: 379663698
Change-Id: Ib5e8ba95cbffd5768de54c4bfdb48fa8d25d3d34
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6051401
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Reviewed-by: Etienne Pierre-Doray <etiennep@chromium.org>
Commit-Queue: Thiabaud Engelbrecht <thiabaud@google.com>
Cr-Commit-Position: refs/heads/main@{#98203}