[js_fuzzer] Soft-skip some more crashtest files
This extends existing logic for soft-skipping files also to regular
expressions checked on the absolute paths. Soft-skipped files will
be chosen with a probability of factor 4 lower.

This is now used to soft-exclude resource files of webgl tests,
which are typically not useful for JS fuzzing. Such files are
among the top offenders for early bailouts atm.

Bug: 389069288
Change-Id: I4050055977fd6fce962d755b923776a5050ee436
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6177727
Reviewed-by: Carl Smith <cffsmith@chromium.org>
Commit-Queue: Michael Achenbach <machenbach@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98211}