Reland "[heap] Fix race in detaching cppheap"
By delaying detaching the cppheap such that it happens after the
backgroudn threads have been finished, it got delayed past the point
where cppheap is able to finish an ongoing GC. With this CL, detaching
the cppheap is split, the part that is finishing an ongoing GC happens
when it happened originally, before background compilation finished. But
the final part of detaching the cppheap only happens at the end of heap
teardown.

This CL also sets up CppHeap in d8, so the tear down code is also
tested.

Original message:

In case the heap owns the cppheap, the cppheap is currently detached as
soon as we start tearing down the heap. At that time background
compilation threads may still be active and may try to allocate, which
means they will check allocation limits and try to access the cppheap to
read the embedder size. This results in a race between detaching the
cppheap and the background compiler.

Fix by delaying detaching the cppheap such that it happens after the
background threads have been stopped.

Bug: 361711106, 42203693
Change-Id: I8459161bf93e06a64406b03bd0b442e6806fe9b1
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6179436
Reviewed-by: Omer Katz <omerkatz@chromium.org>
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98226}