[api] Limit max array buffer size through API.
The embedder can intercept and deny large array buffer allocations through the ArrayBuffer Allocator. If allocations above a certain size are always going to be rejected it makes sense to detect this early and throw exceptions rather than cause a fatal OOM in the VM. With this change the embedder can set an upper limit that will be handled more cleanly.

Change-Id: I3ab54afe7e203d5b82ce9022ae129702fad61c80
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6098133
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Erik Corry <erikcorry@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98290}