[cppgc] Fix how the heap thread is determined
A CppHeap is supposed to be used only by a single thread, the heap
thread. So far, the heap thread was that thread the CppHeap was created
on, or, if the CppHeap was attached to an isolate, the last thread that
acquired the isolate with an Isolate::Locker.

This definition of the heap thread caused problems when the heap thread
changed while the CppHeap was attached to the isolate, but then got
detached from the isolate. Since the CppHeap was not attached to an
isolate anymore, the heap thread was again the thread the CppHeap was
created on, which was incorrect.

With this CL the thread the isolate gets detached on is considered the
new heap thread, and the CppHeap is only allowed to be used on that
thread until it gets terminated, or attached to a new isolate.

R=omerkatz@chromium.org

Bug: 42203693
Change-Id: Ibdc8cf50377eaa4be8bca1621d6fe12f25883005
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6187533
Reviewed-by: Omer Katz <omerkatz@chromium.org>
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98250}