[compiler-dispatcher] Allow QueueForOptimization to fail
With a process-wide OptimizingCompileDispatcher we cannot guarantee
anymore that after a successful `IsQueueAvailable()` the following
`QueueForOptimization()` succeeds as well - other threads might
push into the queue in-between. This CL changes
QueueForOptimization() to allow for failures. The method now requires
a reference to a std::unique_ptr<TurbofanCompilationJob>. It only
takes ownership of the allocation if queuing succeeds and also returns
true.

Bug: 391169065
Change-Id: Ibeffd7dee6607a368817801ee3055f438698baab
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6186007
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98257}