[cppgc][test] Call EnsureSweepingCompleted unconditionally
EnsureSweepingCompleted completes sweeping both for the JS heap and the
CppHeap. However, a condition prevented the call in the case where only
the CppHeap needed to finish sweeping. The condition also exists in
EnsureSweepingCompleted anyways, so it does not change the test to
remove the condition.

R=omerkatz@chromium.org

Bug: 42203693
Change-Id: I6cba49266bda259a8d2a11037959da86a79baae7
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6186582
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Reviewed-by: Omer Katz <omerkatz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98259}