[heap] Avoid repeatedly loading page metadata in css
The young gen conservative visitor loads the the page metadata address
many times per page during page iteration. It only need to the marking
bitmap from the page metadata.
Avoid the repeated loads by stashing the bitmap pointer and passing it
to the visitor as an argument for the methods that need it.

Bug: 379788114
Change-Id: I97512b4195f905f759976a73ab55f62bf1830a89
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6189988
Reviewed-by: Dominik Inführ <dinfuehr@chromium.org>
Commit-Queue: Omer Katz <omerkatz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98273}