[execution] Introduce AddIsolate() and RemoveIsolate() methods
This CL adds new dedicated methods for adding an isolate and removing
an isolate from an IsolateGroup. Both operations are protected by a
mutex, which allows us to make the isolate_count_ in IsolateGroup
non-atomic.

My hope is that this makes it a bit simpler to share stuff between
isolates and having dedicated locations for setting this data up and
destroying it again.

Bug: 391169065
Change-Id: Ia6b39c195b82d1d4ff11b65757169a0d1d3ccc10
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6187326
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98282}