[wasm][compiler] Remove a big chunk of WasmGraphBuilder
Part of it are still used for Wasm-in-JS inlining. Big parts can already
be removed though.

The selection of methods to remove was mostly led by intuition and
coverage data from
https://ci.chromium.org/ui/p/v8/builders/ci/V8%20Linux64%20-%20coverage.
Depending on the coverage after this CL we could run a second iteration
of removal.

R=mliedtke@chromium.org

Bug: 391378125
Change-Id: Id6dddcd6b5d265389e9b9b334d4d8c7e2148292c
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6190459
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Reviewed-by: Matthias Liedtke <mliedtke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98294}