[inspector] Add resolved breakpoints to Debugger.scriptParsed event
The only method in inspector that emits "breakpointResolved" events, is
the same method that emits the "scriptParsed" events.

We can include the resolved breakpoints directly in the "scriptParsed"
event. This way, clients don't have to wait for an indefinite time to
figure out whether they need to re-set breakpoints for a particular
script or not.

We still emit the "breakpointResolved" event separately to stay
backwards compatible, but clients are encouraged to use the list from
the "scriptParsed" event.

R=bmeurer@chromium.org

Bug: None
Change-Id: I48e69317d55cbace39a14e809579533d8dc73aa1
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6198358
Commit-Queue: Simon Zünd <szuend@chromium.org>
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98319}