[collections] Alternative fix for supporting big WeakMaps
The hashes stored in JSReceivers are maximally 20
bits. If a hash table is bigger than that, all
items get clumped in the beginning.

This CL distributes the probes more evenly in big
hash tables. It's still not optimal (there will be
empty space because of quadratic probing) but it's
better than the previous solution.

Side fix: "are we running out of capacity" computation
in RehashObjectHashTableAndGCIfNeeded was too pessimistic
(it added slack twice, first there and once in
ComputeCapacity).

The previous version
https://chromium-review.googlesource.com/c/v8/v8/+/5961516
was applying the hash spreading logic to all HashTables,
while it's only intended for EphemeronHashTable. This CL
fixes that.

Bug: 42207427, 375033082, 390568195
Change-Id: I78dd3f904554fbbe224e4406e79ccb86e7cd7516
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6187094
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Marja Hölttä <marja@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98303}