[wasm][type-reflection] Fix DCHECK in table.grow
When we already have a compiled wrapper for a WasmJSFunction that
is put into a table, then we don't allocate a fresh WasmImportData
for it just to store the table as call origin, because we won't
ever need that call origin any more. So when growing such a table,
entries in the old table might not have a call origin at all (which
was correctly handled), or might have an "incorrect" call origin
(the funcref itself, rather than the table). Either is fine.

Fixed: 390675703
Change-Id: I823a10034013f5d7333b0e9558c1e1e6c0710106
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6197940
Commit-Queue: Matthias Liedtke <mliedtke@chromium.org>
Auto-Submit: Jakob Kummerow <jkummerow@chromium.org>
Reviewed-by: Matthias Liedtke <mliedtke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98305}