[handles] Introduce and DirectHandle::From{Address,Slot}
This CL introduces two static methods, DirectHandle::FromAddress and
DirectHandle::FromSlot, for creating direct handles given an Address
and an Address*, respectively. The former is to be used instead of the
explicit constructor, which is now made private. The latter is to be
used for avoiding implicit IndirectHandle to DirectHandle conversions,
when a slot is given.

Bug: 42203211
Change-Id: I501dc9b3ec84f3cefa3d2ee841ab2ef3858307f6
Cq-Include-Trybots: luci.v8.try:v8_linux64_css_dbg
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6198998
Reviewed-by: Omer Katz <omerkatz@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Nikolaos Papaspyrou <nikolaos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98410}