[flags] Add option to print JS feature flags
This CL adds the --print-js-feature-flags-json, which prints all JS
feature flags as a JSON literal, categorized by in-progress, staged, and
shipped.

This JSON value is intended to be used by scripts to clean up flags in
test files in accordance with guidelines in issue 391658080.

Bug: 391658080
Change-Id: I714c9efa074a85f640a8ffd35b52b4e7d2a7f122
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6199962
Commit-Queue: Shu-yu Guo <syg@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98368}