[turboshaft] Mark receiver/function as escaping in StringEscapeAnalysis
This is similar to https://crrev.com/c/3584117.

The idea is that we iterate the FrameState not only for
deoptimizations (where dematerialized objects are fine), but also when
throwing exceptions, where dematerialized objects are forbidden (but,
in that case we don't look at all the values but only Function and
Receiver).

Fixed: chromium:391500839
Change-Id: I6350a28429a7ac09746422004d999b73bcdec159
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6198350
Auto-Submit: Darius Mercadier <dmercadier@chromium.org>
Commit-Queue: Darius Mercadier <dmercadier@chromium.org>
Reviewed-by: Nico Hartmann <nicohartmann@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98355}