[turboshaft][arm] Port Arm ISEL unittest (part 1)
This CL adds the skeleton for the Arm ISEL unittest: it's just a
copy-paste of the Turbofan one, entirely disabled for now. The next CL
will enable it and actually port it to Turboshaft.

(doing it in 2 CLs makes reviewing easier)

Change-Id: I35f11ad25e57d6d91784fd8cabff48b604964dde
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6207191
Reviewed-by: Matthias Liedtke <mliedtke@chromium.org>
Commit-Queue: Matthias Liedtke <mliedtke@chromium.org>
Auto-Submit: Darius Mercadier <dmercadier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98367}