Add omitted Oilpan size to heap snapshots
When computing the total size of a heap snapshot, it would be useful to
know how many bytes of Oilpan objects were either omitted or represented
as occupying zero size. This will enable DevTools to compute a total
size for the snapshot that better matches the total size shown in the
"Select JavaScript VM instance" list.

Bug: 40734351
Change-Id: I2afb5393ffcb36dcd3fd6b53f4506c692fa82f86
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6196637
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Seth Brenith <seth.brenith@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#98390}