[turboshaft][isel] Fix branch-if-overflow fusion
The logic for branch-if-overflow in the ISEL currently expects:

    1: TheOverflowingOperation
    2: OverflowProjection(1)
    ...
    3: Branch(2)

    <later block>:
    4: ValueProjection(1)
    // uses of 4

And the branch-if-overflow fusion logic would check that when visiting
`3: Branch(2)`, the value projection would already be defined.

In Turbofan, projections used to be scheduled late (apparently), which
means that such graphs would be possible. However, in Turboshaft,
projections are always scheduled eagerly, so we would have something
like that (keeping the same instruction numbers as Turbofan, for
clarity):

    1: TheOverflowingOperation
    4: ValueProjection(1)
    2: OverflowProjection(1)
    ...
    3: Branch(2)

    <later block>:
    // uses of 4

Which means that when visiting `3: Branch(2)`, the value projection
can never be already defined, and the branch-if-overflow fusion can
never happen.

This CL fixes that by changing what we check to do the
branch-if-overflow: instead of checking that the projection is
defined, we check that all of its uses are either defined or are in
other (later) blocks.

Change-Id: Ia46c4dd2fe164d2dc44c067dec6a782c10823294
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6206471
Commit-Queue: Darius Mercadier <dmercadier@chromium.org>
Reviewed-by: Matthias Liedtke <mliedtke@chromium.org>
Auto-Submit: Darius Mercadier <dmercadier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98387}