[js_fuzzer] Reduce test input from in-progress features
Certain flags are dropped from fuzzing due to false positives, e.g.
during the in-progress state of a new language feature. However,
using test cases that depend on such flags as input, leads to poor
fuzzer performance.

This change now drops such files with a high probability. To make
this decision, we need the unfiltered, original flags from the
input test cases. Therefore, the first patchsets refactor the code
and move the filtering of flags to the end of the fuzzing.

Patch 1 adds an additional test that covers the flag
filtering output of normal testing:
https://crrev.com/c/6216922/1

Patch 2 drops the early filtering and demonstrates (running the
trybot) how tests break:
https://crrev.com/c/6216922/1..2

Patch 3 fixes tests again by moving the filtering towards the end
of the fuzzing process.
https://crrev.com/c/6216922/2..3

The remaining patches implement the feature, dropping files based
on their flags contents:
https://crrev.com/c/6216922/3..5

This adds the in-progress flags and a few other flags, whose
respective files are not useful, to a separate config list.

The mjsunit files were previously tidied up as part of bug 391658080
so that no wrong flags are left.

Bug: 391658080
Change-Id: I7bd69cefa036d51a43f17df38c756c4ffff0bb83
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6216922
Reviewed-by: Stephen Röttger <sroettger@google.com>
Commit-Queue: Michael Achenbach <machenbach@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98427}