[js_fuzzer] Tune try catch
The try-catch mutator runs always and wraps statements with
try-catch. It also drops the wrapping with a probability based on
the location in the sources (higher prob towards the end) and also
might drop nested try-catches.

The probabilities were at some point arbitrarily chosen. Data from
experiments shows that we bail out too often due to missing
try-catch wrapping on the top level. At the same time, we could
do with less wrapping on nested levels.

Hence, this changes the probabilities to drop fewer try-catches.
The probability for dropping nested try-catch is now not depending
on the source location any longer, but equal across the file.

Because the effects are rather difficult to assess, this adds
another test case with a large expectation file that exemplifies
the distribution. The test is pseudo random.

Patch 0..3 adds the new test and shows the distribution with the
status quo:
https://chromium-review.googlesource.com/c/v8/v8/+/6218497/3

Patch 4 adds the tuning and shows the resulting test changes:
https://crrev.com/c/6218497/3..4

The test shows now more wrapping at the top-level, while dropping
a few more nested wrappings.

In a follow up we might also take existing try-catch into account,
in which case we could skip some more superfluous wrapping.

Bug: 389069288
Change-Id: I674500265be5457a2d57d8677571598453d773e5
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6218497
Reviewed-by: Carl Smith <cffsmith@chromium.org>
Commit-Queue: Michael Achenbach <machenbach@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98471}