[float16array] Move float16->float64 into representation changer
The current way the graph is assembled for LoadTypedElement on
Float16Arrays is to manually wrap the load in a Float16RawBitsToNumber
node. This is error prone as it requires the producer of
LoadTypedElement to know Float16Arrays need special handling. It also
causes mismatches the type of LoadTypedElement, which is Number for
Float16Arrays and causes --assert-types to assert.

This CL removes the manual wrapping and adds support in the
representation changer for float16 raw bits conversion. These
conversions have a Word16 machine representation and the Number type.
They are only emitted when the opcode is LoadTypedElement or
LoadDataViewElement.

Bug: 42203953
Change-Id: I5ff4a6d9f702d6f6a84489a1ee890da5279380d8
Fixed: 392661402
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6222241
Reviewed-by: Darius Mercadier <dmercadier@chromium.org>
Commit-Queue: Shu-yu Guo <syg@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98478}