[js_fuzzer] Prevent syntax errors with try-catch wrapping and await
The try-catch wrapper wraps code with an anonymous arrow function.
If the wrapped code is an await expression, it can't remain in
this function not marked as async, as this produces syntax errors.

Patch 1 shows the status quo with some new test data:
https://crrev.com/c/6225240/1

Patch 2 fixes the problem by pulling await keywords outside of the
wrapping expression:
https://crrev.com/c/6225240/1..2

Bug: 389069288
Change-Id: I64cb195c734273961ee2772b5443cc6428069e87
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6225240
Reviewed-by: Carl Smith <cffsmith@chromium.org>
Commit-Queue: Michael Achenbach <machenbach@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98473}