[heap] Remove redundant marking bitmap check
The attached bug is a race between checking that the bitmap is clear and
later setting a bit in the bitmap during concurrent marking. The check
is redundant since we already have the same check in line 47 above, that
is called when allocating the page and is guarded by a fence.
Fix by dropping the duplicate check in line 150.

Bug: 394052225
Change-Id: Ibb0404fdd48733668b2fd7ba3ce9d268cd675066
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6227880
Auto-Submit: Omer Katz <omerkatz@chromium.org>
Commit-Queue: Michael Lippautz <mlippautz@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98485}