[maglev] Better recording of Phi in-loop use hints
The previous approach to figuring out if a use was inside a loop or
not was broken because of inlining, because `current_offset` refers to
the bytecode offset of the current innermost function being inlined
rather than the offset in the function that created the loop phi. So,
consider this small example:

    function bar(x) {
      return x + 2; // bytecode offset = 0
    }

    function foo() {
      // some random stuff
      let phi = 0;
      for (let i = 0; i < 42; i++) { // bytecode offset = 4
        phi = bar(phi); // bytecode offset = 6
      } // bytecode offset = 8
      return phi; // Tagged use
    }

When `bar` is inlined, this creates a Int32 use of `phi`, but at
offset 0 (since that's the bytecode offset in the inlinee), which it
outside of the range for `phi`'s loop (which is from offset 4 to
offset 8), and thus it wasn't being recorded as a use inside the loop.

When now consider that all uses are inside of the loop until the
backedge has been merged.

Bug: 393335991
Change-Id: I61f6f620e40212a3fcc113da370a7df0629faae1
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6228060
Auto-Submit: Darius Mercadier <dmercadier@chromium.org>
Commit-Queue: Olivier Flückiger <olivf@chromium.org>
Reviewed-by: Olivier Flückiger <olivf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98487}