[js_fuzzer] Prevent duplicate variable declaration through cross-over insertion
Snippets depending on two variable names might just declare one in
the scope of the other, which leads to syntax errors.

This change ensures that we choose distinct variables if more than
one is required. This also replaces an old inefficient
implementation for random samples for iterables. We currently only
need samples of 2 from arrays.

Patch 2 adds a test scenario that shows the status quo. The test
contains a fake snippet DB with one snippet depending on two
variables. The pseudo-random output wrongly choose the same
variable twice for one of the insertions:
https://crrev.com/c/6226079/2

Patch 6 implements the fix and also replaces an old implementation
of sample with a simpler and faster one.
https://crrev.com/c/6226079/2..6

Bug: 389069288
Change-Id: I66d20049567a616f81846b2abc4c3dc89146474d
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6226079
Reviewed-by: Stephen Röttger <sroettger@google.com>
Commit-Queue: Michael Achenbach <machenbach@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98544}