[liftoff] Fix memory bounds check to not clobber tagged stack slots
in debugging mode.

When debugging mode is enabled, liftoff spills temporary registers.
The tagged state of registers and stack slots is snapshotted when
creating the out of line label. Afterwards, no more registers may be
requested / allocated as this could cause spilling of values from the
value stack and by that invalidate the taggedness of registers and
stack slots of the out of line label.

Fixed: 391921972
Change-Id: I88f7b09283f2985cd1a7e9bade7cd10c61af8f96
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6227566
Commit-Queue: Matthias Liedtke <mliedtke@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98496}