[turbofan] Refactor: Rename Safe to Small in speculative additions
This commit renames the Safe terminology to Small in the context of
speculative integer additions. The previous naming suggested that the
additions were performed within the "safe" integer range, which was not
the case. This change clarifies that the speculative additions assume
the operands are small integers, as indicated by the
NumberOperationHint::kSignedSmall hint used in
VisitSpeculativeSafeIntegerAdditiveOp.

turboshaft as an speculative operation.

Drive-by: We should never lower SpeculativeSafe/SmallIntegerAdd to
Change-Id: I46991a242472b4a45873d62eedfc2fa0c415ded0
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6233893
Commit-Queue: Nico Hartmann <nicohartmann@chromium.org>
Auto-Submit: Victor Gomes <victorgomes@chromium.org>
Reviewed-by: Nico Hartmann <nicohartmann@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98516}