[json] Stringfy: Front-load some checks for fast-path
We need to check that we don't have interesting properties (toJSON)
anywhere on JSObject and JSArray prototype chains.
Previously this was ensured whenever we serialize a JSObject/JSArray,
but we can front-load all the checks on the prototypes to the start of
serialization and only check that the prototype is the expected one for
each object.

Bug: 380044242
Change-Id: I2f736602c72b3e4045fbbcbb99191fe6d0f97343
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6233933
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Commit-Queue: Patrick Thier <pthier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98687}