[wasm][type-reflection] Do not unwrap WasmJSFunctions on import
Use the WasmJSFunction object itself as the callable target for imports
and table entries, and not the callable object that it holds.
Otherwise the inner object could be another WasmJSFunction with a
different (invalid) signature. This object can be re-validated again for
linking when the generic wrapper tiers up, and we assume that the
signature is correct at this point.

R=jkummerow@chromium.org
CC=​irezvov@chromium.org

Fixed: 393632542
Change-Id: I1db849af5f3cddf0daf5dae4142071f415897ac7
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6225722
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Thibaud Michaud <thibaudm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98515}