[string] Optimize and refactor index hashing
gotos are considered harmful and switches are not, so extract out
helper functions from index hashing which report whether the index
contained non-index (non-digit or leading zero) values, or whether it
simply overflowed.

Also do some microoptimisation that merges the index check with the
conversion to decimal, and changes the overflow check to detect overflow
(with a single branch) rather than prevent it (with branches dependent
on both the current digit and the value so far). This should reduce
branches and tighten the loop.

Change-Id: Id279a6b6a74c99316cb99b48bacbc60679605bab
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6218917
Commit-Queue: Leszek Swirski <leszeks@chromium.org>
Reviewed-by: Patrick Thier <pthier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98519}