[js_fuzzer] Reduce cross-async function replacements
Exchanging calls to async functions with sync functions and the other
way around leads to bail outs and inferior test cases.

This change maintains the async property during function replacement
with an exception of a small probability. As a fly-by fix, this also
prevents replacements with the same function.

In detail:
- This adds a new context class that can store extra state throughout
one fuzz-test generation. Previously there was no way to do this. In
the future, this can be extended with other things that are
currently implemented a bit hacky (e.g. extra state glued to the AST
objects).
- The context collects now the set of declared async functions during
the identifier normalization phase that runs before mutation.
- The function-call mutator uses this information when making
function-call replacements.

Bug: 389069288
Change-Id: I096380ab10fc0e90d4d928ae580bb44649b8673f
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6227578
Reviewed-by: Stephen Röttger <sroettger@google.com>
Commit-Queue: Michael Achenbach <machenbach@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98545}