[wasm] Verify max operation input count in the wasm-to-js wrapper
JS has a limit of 2^16-1 parameters per function, but the compiled
wasm-to-js wrapper adds a few implicit arguments to the call operation.
This can create an op with more than 2^16-1 inputs, which breaks
Turboshaft's implementation limit.
Temporarily turn this into a hard crash with a check, and follow up with
a more robust fix.

R=jkummerow@chromium.org

Bug: 394350433
Change-Id: I334a26053c3aa4c55e7f26d7fa228689c0587308
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6239200
Commit-Queue: Thibaud Michaud <thibaudm@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98541}