[HeapSnapshot] Serialize numbers directly into the snapshot chunk
A large part of HeapSnapshotJSONSerializer's work is serializing numbers
into the JSON snapshot chunk. The current implementation uses bespoke
logic to convert numbers to strings in buffers and MemCopy's the buffers
into the chunk.

std::to_chars()[1] is an API introduced in C++17 to convert numbers to a
string directly into a supplied char buffer without additional memory
allocation. This CL replaces the bespoke logic with std::to_chars() and
gets rid of the intermediate buffers. The benefits are:
- No additional memory allocation for the intermediate buffers.
- Better memory safety through use of standard library functions.
- Simplified code that handles all types of numbers.

As noted in the Faster Transfer of Heap Snapshots doc[2], calls to
strlen() contribute to ~7% of the total time spent in serializing the
snapshot. This CL also gets rid of the majority of these calls and thus
also speeds up snapshot serialization. On a Linux x64 machine, the
average serialization time for a snapshot of YouTube.com (85.3 MB, 1.89
million nodes, 10.5 million edges) reduced from 3800 ms to 3580 ms,
which is a ~6% improvement.

[1] https://en.cppreference.com/w/cpp/utility/to_chars
[2] https://docs.google.com/document/d/1o5sOSWe8H68X7Zn8zPqpF-bo5rAj410e9vvUFZ_NU0s/edit?usp=sharing

Change-Id: I6b888fb0bc572d8633d4557f379f6d6ff1a8ba49
Bug: 387077340
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6217633
Reviewed-by: Seth Brenith <seth.brenith@microsoft.com>
Commit-Queue: Abhishek Shanthkumar <abhishek.shanthkumar@microsoft.com>
Reviewed-by: Simon Zünd <szuend@chromium.org>
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98576}