[turboshaft] Fix wrong FrameState.data comparison in DCHECK
Since https://crrev.com/c/6226679, FrameStates can be GVNed. As a
result, if we have something like

    B1 (loop):
      v1: FrameState(...)
      ...
      Branch(c, B2, B3)

   B2:
      v2: FrameState(...)
      Goto(B1)

   B3: (after loop)

If we unroll the loop, and the `v1` from the second iteration can be
GVNed by the `v2` from the previous iteration, then this new `v1` will
not have the same FrameStateData pointer as the one from the first
iteration. And, when emitting `B3`, it will now have two predecessors,
and will thus try to merge all variables, including the two versions
of `v1`.
This is perfectly safe, and the DCHECK in VariableReducer
just needs to be adapter to not compare pointers, but values instead.

Fixed: 394402574
Change-Id: I6b9fac5ee677012f3ebb4706c6f955fa6187eded
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6233649
Auto-Submit: Darius Mercadier <dmercadier@chromium.org>
Commit-Queue: Nico Hartmann <nicohartmann@chromium.org>
Reviewed-by: Nico Hartmann <nicohartmann@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98550}